Increase the continuous error statistics, automatically stop more than three times
