+++ xmrminer/src/main/java/com/cpf/xmrminer/Job.kt
@@ -2,7 +2,6 @@ package com.cpf.xmrminer
 
 import com.cpf.baseproject.util.logPrint
 import com.cpf.cryptonight.Miner
-import org.json.JSONObject
 
 /**
  * Created by cpf on 2017/12/28.
@@ -28,9 +27,7 @@ class Job(private val miner: MinerManager,
             Miner.fastHash(blobByte, result)
             val resultStr = HexUtil.hexlify(result)
             val rs = resultStr.substring(resultStr.length - jobBean.target.length)
-            var byteArray = HexUtil.unhexlify(rs)
-            byteArray = miner.byteArraySort(byteArray)
-            val str = HexUtil.hexlify(byteArray)
+            val str = miner.swapEndian(rs)
             val tarL = str.toLong(16)
             if (tarL < jobBean.target.toLong(16)) {
                 logPrint("submit____id${miner.mId}___thread${start}_____${jobBean.target}_____${str}_____$resultStr")
@@ -38,7 +35,7 @@ class Job(private val miner: MinerManager,
                 nonceByte[1] = (nonce shr 8).toByte()
                 nonceByte[2] = (nonce shr 16).toByte()
                 nonceByte[3] = (nonce shr 24).toByte()
-                send(HexUtil.hexlify(nonceByte), resultStr)
+                miner.send(jobBean, HexUtil.hexlify(nonceByte), resultStr)
             }
             nonce += space
             miner.mHashCount++
@@ -49,21 +46,6 @@ class Job(private val miner: MinerManager,
                 }
             }
         }
-    }
-
-    private fun send(nonceHex: String, tar: String) {
-        val jsonObj = JSONObject()
-        jsonObj.put("id", miner.mId++)
-        jsonObj.put("method", "submit")
-        jsonObj.put("jsonrpc", "2.0")
-        val json = JSONObject()
-        json.put("id", jobBean.id)
-        json.put("job_id", jobBean.jobId)
-        json.put("nonce", nonceHex)
-        json.put("result", tar)
-        jsonObj.put("params", json)
-        miner.mPrintWrite?.println(jsonObj.toString())
-        miner.mPrintWrite?.flush()
     }
 
     class JobBean {
